Divide settings.jsx into components
